backend/requirements.txt
fastapi
python-multipart
pillow
numpy
opencv-python-headless
uvicorn